+++ src/Dialog.svelte
@@ -40,7 +40,6 @@
 
 	export { className as class, style, visible, width };
 
-	/* eslint-disable no-unused-vars */
 	let className = '';
 	let style = '';
 	let visible = false;
@@ -49,6 +48,7 @@
 	let attrs = {};
 
 	$: {
+		/* eslint-disable no-unused-vars */
 		const { style, visible, width, ...other } = $$props;
 
 		attrs = other;
@@ -122,6 +122,7 @@
 		position: relative;
 		font-size: 1rem;
 		background: #eee;
+		/* postcss-custom-properties: ignore next */
 		background: var(--bg-panel, #eee);
 		border-radius: 4px;
 		cursor: auto;
@@ -162,6 +163,7 @@
 		font-size: 24px;
 		line-height: 36px;
 		background: rgba(0, 0, 0, 0.1);
+		/* postcss-custom-properties: ignore next */
 		background: var(--divider, rgba(0, 0, 0, 0.1));
 	}
 	.content {