+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unpablished
 
 - Menu
+	- use popover component
 	- add class, style props
 
 ## 0.1.4